Update Usecase Tinh tong so luong giao dich theo tưng loai
